requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
pytesseract>=0.3.10

# Web automation
selenium>=4.8.0
webdriver-manager>=3.8.0

# Image processing
opencv-python>=4.7.0
numpy>=1.21.0
scipy>=1.9.0
scikit-image>=0.19.0

# Utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.64.0
colorama>=0.4.6

# Optional: Machine learning for advanced CAPTCHA solving
# tensorflow>=2.11.0
# torch>=1.13.0
# torchvision>=0.14.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0